+++ lighthouse/entrypoint.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 cd /etc/monitor-worker/lighthouse/
-npm i
+npm i -g  --unsafe-perm=true --allow-root
 npm start